requirements.txt
pygame==2.0.0.dev3
pyserial
